Temporarily pin meson to 0.55.3 as latest 0.56.0 is failing
The problem appears to be that libraries listed in
Requires.private field of dependencies' pkg-config .pc files
are ignored when building meson subproject executables.

(We don't really care about these executables, but most
dependencies don't offer the option to skip building them.)

There are quite a few dependency-related changes in meson 0.56.0
so it's not obvious where this has broken.